Fix SubChunkRequestPacket deserialization

+++ bedrock/bedrock-common/src/main/java/com/nukkitx/protocol/bedrock/packet/SubChunkRequestPacket.java
@@ -18,7 +18,7 @@ public class SubChunkRequestPacket extends BedrockPacket {
     /**
      * @since v485
      */
-    private int requestCount;
+    private long requestCount;
     /**
      * @since v485
      */
